Made some adjustments in the code.

+++ src/org/hsm/view/gui/View.java
@@ -45,7 +45,7 @@ public interface View extends VisibleComponent {
      * @param conductivity the conductivity of the plant
      * @param temperature the temperature of the plant
      */
-    void insertNewPlant(final int id, final String name, final double cost, final int ph, final double brightness,
+    void insertNewPlant(final int id, final String name, final double cost, final double ph, final double brightness,
              final double conductivity, final double temperature);
 
     /**
